add spouting spirits
