+++ src/interface/others/RestorationShamanSpreadsheet.js
@@ -49,7 +49,7 @@ class RestorationShamanSpreadsheet extends React.Component {
               <tr><td>{parser.selectedCombatant.hasTalent(SPELLS.CLOUDBURST_TOTEM_TALENT.id) ? cpm(SPELLS.CLOUDBURST_TOTEM_TALENT.id) : cpm(SPELLS.HEALING_STREAM_TOTEM_CAST.id)}</td></tr>
               <tr><td>{cpm(SPELLS.CHAIN_HEAL.id)}</td></tr>
               <tr><td>{parser._modules.spreadsheet.surgingTideProcsPerMinute}</td></tr>
-              <tr><td>{(parser._modules.spreadsheet.spoutingSpiritsHits / casts(SPELLS.SPIRIT_LINK_TOTEM.id) || 0).toFixed(2)}</td></tr>
+              <tr><td>{(parser._modules.spoutingSpirits.spoutingSpiritsHits / casts(SPELLS.SPIRIT_LINK_TOTEM.id) || getAbility(SPELLS.SPOUTING_SPIRITS_HEAL.id).healingHits / casts(SPELLS.SPIRIT_LINK_TOTEM.id) || 0).toFixed(2)}</td></tr>
               <tr><td>{(parser._modules.overflowingShores.overflowingShoresHits / casts(SPELLS.HEALING_RAIN_CAST.id) || 0).toFixed(2)}</td></tr>
               <tr><td>{parser._modules.ebbAndFlow.ebbAndFlowEffectiveness.toFixed(2)}</td></tr>
               <tr><td>{parser._modules.combatants.playerCount}</td></tr>
